Store Prose View in meta thought.
