+++ src/shortcuts/index.js
@@ -19,6 +19,7 @@ import newThoughtAbove from './newThoughtAbove.js'
 import newUncle from './newUncle.js'
 import openShortcutPopup from './openShortcutPopup.js'
 import outdent from './outdent.js'
+import proseView from './proseView.js'
 import redo from './redo.js'
 import search from './search.js'
 import subcategorizeAll from './subcategorizeAll.js'
@@ -26,7 +27,6 @@ import subcategorizeOne from './subcategorizeOne.js'
 import toggleCodeView from './toggleCodeView.js'
 import toggleContextView from './toggleContextView.js'
 import toggleHiddenThoughts from './toggleHiddenThoughts.js'
-import toggleProseView from './toggleProseView.js'
 import toggleSplitView from './toggleSplitView.js'
 import toggleTableView from './toggleTableView.js'
 import undo from './undo.js'
@@ -56,6 +56,7 @@ export {
   newUncle,
   openShortcutPopup,
   outdent,
+  proseView,
   redo,
   search,
   subcategorizeAll,
@@ -63,7 +64,6 @@ export {
   toggleCodeView,
   toggleContextView,
   toggleHiddenThoughts,
-  toggleProseView,
   toggleSplitView,
   toggleTableView,
   undo,
